Convert db to EF Core; start on view page
Also returning skills with profile inquiries now, though that particular query is failing (current WIP) #2

+++ src/JobsJobsJobs/Shared/Api/Count.cs
@@ -3,5 +3,5 @@
     /// <summary>
     /// A transport mechanism to send counts across the wire via JSON
     /// </summary>
-    public record Count(long Value);
+    public record Count(int Value);
 }
